+++ app/src/main/java/com/example/proyectoZapateria/data/model/UsuarioCompleto.kt
@@ -2,19 +2,19 @@ package com.example.proyectoZapateria.data.model
 
 // Modelo unificado que representa un usuario completo combinando datos de Persona, Usuario y Rol
 data class UsuarioCompleto(
-    val idPersona: Int,
+    val idPersona: Long,
     val nombre: String,
     val apellido: String,
     val rut: String,
     val telefono: String,
     val email: String,
-    val idComuna: Int?,
+    val idComuna: Long?,
     val calle: String,
     val numeroPuerta: String,
     val username: String,
     val fechaRegistro: Long,
     val estado: String,
-    val idRol: Int,
+    val idRol: Long,
     val nombreRol: String,
     val activo: Boolean = true
 )
