+++ app/src/main/java/com/example/proyectoZapateria/data/remote/entregas/EntregasApiService.kt
@@ -10,31 +10,29 @@ interface EntregasApiService {
     suspend fun obtenerTodasLasEntregas(): Response<List<EntregaDTO>>
 
     @GET("api/entregas/{id}")
-    suspend fun obtenerEntregaPorId(@Path("id") id: Int): Response<EntregaDTO>
+    suspend fun obtenerEntregaPorId(@Path("id") id: Long): Response<EntregaDTO>
 
     @GET("api/entregas/transportista/{transportistaId}")
-    suspend fun obtenerEntregasPorTransportista(@Path("transportistaId") transportistaId: Int): Response<List<EntregaDTO>>
+    suspend fun obtenerEntregasPorTransportista(@Path("transportistaId") transportistaId: Long): Response<List<EntregaDTO>>
 
     @GET("api/entregas/estado/{estado}")
     suspend fun obtenerEntregasPorEstado(@Path("estado") estado: String): Response<List<EntregaDTO>>
 
     @PUT("api/entregas/{id}/asignar")
     suspend fun asignarTransportista(
-        @Path("id") id: Int,
-        @Query("transportistaId") transportistaId: Int
+        @Path("id") id: Long,
+        @Query("transportistaId") transportistaId: Long
     ): Response<EntregaDTO>
 
     @PUT("api/entregas/{id}/completar")
     suspend fun completarEntrega(
-        @Path("id") id: Int,
+        @Path("id") id: Long,
         @Body observacion: String?
     ): Response<EntregaDTO>
 
     @PUT("api/entregas/{id}/estado")
     suspend fun cambiarEstadoEntrega(
-        @Path("id") id: Int,
+        @Path("id") id: Long,
         @Query("nuevoEstado") nuevoEstado: String
     ): Response<EntregaDTO>
 }
-
-
