+++ app/src/main/java/com/example/proyectoZapateria/data/remote/entregas/dto/EntregaDTO.kt
@@ -1,9 +1,9 @@
 package com.example.proyectoZapateria.data.remote.entregas.dto
 
 data class EntregaDTO(
-    val idEntrega: Int? = null,
-    val idBoleta: Int,
-    val idTransportista: Int?,
+    val idEntrega: Long? = null,
+    val idBoleta: Long,
+    val idTransportista: Long?,
     val estadoEntrega: String,
     val fechaAsignacion: String,
     val fechaEntrega: String?,
@@ -12,4 +12,3 @@ data class EntregaDTO(
     val nombreCliente: String? = null,
     val direccionEntrega: String? = null
 )
-
