+++ app/src/main/java/com/example/proyectoZapateria/data/remote/ventas/dto/BoletaDTO.kt
@@ -1,8 +1,8 @@
 package com.example.proyectoZapateria.data.remote.ventas.dto
 
 data class BoletaDTO(
-    val id: Int? = null,
-    val clienteId: Int,
+    val id: Long? = null,
+    val clienteId: Long,
     val fechaVenta: String,
     val total: Int,
     val estado: String,
