+++ app/src/main/java/com/example/proyectoZapateria/data/remote/ventas/dto/DetalleBoletaDTO.kt
@@ -1,13 +1,12 @@
 package com.example.proyectoZapateria.data.remote.ventas.dto
 
 data class DetalleBoletaDTO(
-    val id: Int? = null,
-    val boletaId: Int? = null,
-    val inventarioId: Int,
+    val id: Long? = null,
+    val boletaId: Long? = null,
+    val inventarioId: Long,
     val nombreProducto: String? = null,
     val talla: String? = null,
     val cantidad: Int,
     val precioUnitario: Int,
     val subtotal: Int
 )
-
